requirements.txt
flask==2.3.3
boto3==1.34.0
python-dotenv==1.0.0
waitress==2.1.2
